refactor(lsp): moved CompletionPriority.scala into completion directory
