+++ main/src/ca/uwaterloo/flix/api/lsp/provider/completion/TypeBuiltinCompleter.scala
@@ -17,7 +17,6 @@ package ca.uwaterloo.flix.api.lsp.provider.completion
 
 import ca.uwaterloo.flix.api.Flix
 import ca.uwaterloo.flix.api.lsp.{Index, InsertTextFormat, TextEdit}
-import ca.uwaterloo.flix.api.lsp.provider.CompletionPriority
 import ca.uwaterloo.flix.api.lsp.provider.completion.Completion.TypeBuiltinCompletion
 import ca.uwaterloo.flix.language.ast.TypedAst
